Use abstract classes to make events DRY
