+++ src/SectionField/Event/ApiEntryDeleted.php
@@ -13,10 +13,6 @@ declare(strict_types=1);
 
 namespace Tardigrades\SectionField\Event;
 
-use Symfony\Component\EventDispatcher\Event;
-use Symfony\Component\HttpFoundation\Request;
-use Tardigrades\SectionField\Generator\CommonSectionInterface;
-
 /**
  * Class ApiEntryDeleted
  *
@@ -24,41 +20,7 @@ use Tardigrades\SectionField\Generator\CommonSectionInterface;
  *
  * @package Tardigrades\SectionField\Event
  */
-class ApiEntryDeleted extends Event
+class ApiEntryDeleted extends ApiAfterEntryEvent
 {
     const NAME = 'api.entry.deleted';
-
-    /** @var Request */
-    protected $request;
-
-    /** @var array */
-    protected $response;
-
-    /** @var CommonSectionInterface */
-    protected $entry;
-
-    public function __construct(Request $request, array $response, CommonSectionInterface $entry)
-    {
-        $this->request = $request;
-        $this->response = $response;
-        $this->entry = $entry;
-    }
-
-    /** @return Request */
-    public function getRequest(): Request
-    {
-        return $this->request;
-    }
-
-    /** @return array */
-    public function getResponse(): array
-    {
-        return $this->response;
-    }
-
-    /** @return CommonSectionInterface */
-    public function getEntry(): CommonSectionInterface
-    {
-        return $this->entry;
-    }
 }
